Updated the backend script to be invoked via python to be able to pass in flags. The loginDev has been moved to a dev_router and will only appear during a development deployment, removing the code entirely in production. Tests and routes have been updated
