+++ backend/Dockerfile
@@ -13,4 +13,5 @@ COPY requirements.txt /app/requirements.txt
 RUN pip install --no-cache-dir -r /app/requirements.txt
 COPY ./src /app/src
 COPY etc/entrypoint-init.sh /app/entrypoint-init.sh
+RUN rm /app/src/routers/dev_router.py
 ENTRYPOINT ["/bin/sh", "-c", "uvicorn src.main:app --host 0.0.0.0 --port 8000 --log-level info"]
